hotfix: Migrate from Pipenv to Poetry package management
- Rename src/ to prisma/ to align with Poetry project naming conventions
- Update pyproject.toml with complete Poetry configuration including CLI entry points
- Fix all test import references from 'src.' to 'prisma.'
- Replace Pipfile with poetry.lock for dependency management
- Update VS Code configuration for Poetry virtual environment
- Create prisma CLI entry point using Click framework
- Ensure all 86 tests pass with new package structure
- Update .vscode/tasks.json to use Poetry instead of Pipenv commands

All functionality maintained with improved dependency management and development workflow.
